lesson_26_homework_21

+++ src/main/java/repo/LessonDao.java
@@ -1,6 +1,9 @@
 package repo;
 
 import model.Lesson;
+import my.exceptions.DublicatedHomeWorkIDException;
+import my.exceptions.LessonDoesNotExistsException;
+import my.exceptions.LessonsArrayIsEmptyException;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -17,46 +20,54 @@ public class LessonDao implements DaoAble {
     }
 
     @Override
-    public void addLesson(String name, String date, int homework) {
+    public void addLesson(String name, String date, int homework) throws DublicatedHomeWorkIDException {
+        int resultSet = 0;
         try (PreparedStatement statement = connection.prepareStatement("INSERT INTO test_mysql_db.Lesson (name, updatedAt, homework_id)" +
                 "VALUES (?, ?, ?)")) {
             statement.setString(1, name);
             statement.setString(2, date);
             statement.setInt(3, homework);
             statement.executeUpdate();
-            int resultSet = statement.getUpdateCount();
+            resultSet = statement.getUpdateCount();
 
             if(resultSet > 0)
                 System.out.println("Lesson was added successfully: " + resultSet);
             System.out.println("Last added lesson id is: " + getLastAddedLessonId() + "\n");
         } catch (SQLException e) {
-            throw new RuntimeException(e);
+            if(resultSet == 0)
+                throw new DublicatedHomeWorkIDException("Homework with id: " + homework + " is already exists. " +
+                        "Specified homeworkID should be unique.");
         }
     }
 
     @Override
-    public void deleteLesson() {
-        try (PreparedStatement statement = connection.prepareStatement("DELETE FROM test_mysql_db.lesson WHERE id = ?")) {
-            statement.setInt(1, getLastAddedLessonId());
-            int rowsDeleted = statement.executeUpdate();
+    public void deleteLesson(int id) throws LessonDoesNotExistsException {
+        if (isLessonPresent(id)){
+            int rowsDeleted = -1;
+            try (PreparedStatement statement = connection.prepareStatement("DELETE FROM test_mysql_db.lesson WHERE id = ?")) {
+                statement.setInt(1, id);
+                rowsDeleted = statement.executeUpdate();
 
-            if(rowsDeleted >= 1)
-                System.out.println("Rows deleted: " + rowsDeleted);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+                if (rowsDeleted >= 1)
+                    System.out.println("Rows deleted: " + rowsDeleted);
+            } catch (SQLException throwables) {
+                throwables.printStackTrace();
+            }
+        } else
+            throw new LessonDoesNotExistsException("Lesson with id \" + id + \" does not exists");
     }
 
     @Override
-    public List<Lesson> getAllLessons() {
+    public List<Lesson> getAllLessons() throws LessonsArrayIsEmptyException {
         String sql = "SELECT * FROM test_mysql_db.lesson";
         List<Lesson> lessons = new ArrayList<>();
+        Lesson lesson;
 
         try (PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.execute();
             ResultSet resultSet = statement.getResultSet();
 
-            Lesson lesson;
+
 
             while (resultSet.next()) {
                     lesson = new Lesson();
@@ -67,16 +78,18 @@ public class LessonDao implements DaoAble {
                     lessons.add(lesson);
             }
 
+            if(lessons.size() == 0)
+                throw new LessonsArrayIsEmptyException("Lessons array is empty.");
+
             System.out.println(lessons);
             return lessons;
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
         }
-        return null;
     }
 
     @Override
-    public Lesson getLessonById(int id) {
+    public Lesson getLessonById(int id) throws LessonDoesNotExistsException {
 
         try (PreparedStatement statement = connection.prepareStatement("SELECT * FROM test_mysql_db.lesson WHERE id = ?")) {
             statement.setInt(1, id);
@@ -91,12 +104,16 @@ public class LessonDao implements DaoAble {
 
             System.out.printf("Lesson with id %d is: " + lesson + "\n", id);
             return lesson;
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+        }
+        catch (SQLException e) {
+            throw new LessonDoesNotExistsException("Lesson with id \"" + id + "\" doesn't exists. " +
+                    "SQLState is: " + e.getSQLState() +
+                    ". SQLErrorCode is " + e.getErrorCode());
+
         }
     }
 
-    private int getLastAddedLessonId(){
+    public int getLastAddedLessonId() throws LessonDoesNotExistsException {
         try (PreparedStatement statement = connection.prepareStatement("SELECT id FROM test_mysql_db.lesson ORDER BY id DESC LIMIT 1")) {
             statement.execute();
             ResultSet resultSet = statement.getResultSet();
@@ -104,7 +121,22 @@ public class LessonDao implements DaoAble {
 
             return resultSet.getInt("id");
         } catch (SQLException e) {
-            throw new RuntimeException(e);
+            throw new LessonDoesNotExistsException("Lessons table is empty.");
         }
+    }
+
+    public boolean isLessonPresent(int id) {
+        try (PreparedStatement statement = connection.prepareStatement("SELECT * FROM test_mysql_db.lesson WHERE id = ?")) {
+            statement.setInt(1, id);
+            statement.execute();
+            ResultSet resultSet = statement.getResultSet();
+
+            if(resultSet.next()){
+                return true;
+            }
+        } catch (SQLException throwables) {
+            throwables.printStackTrace();
+        }
+        return false;
     }
 }
